rename minor cleanup.
