+++ test/com/yetu/oauth2provider/base/BaseMethods.scala
@@ -1,8 +1,8 @@
 package com.yetu.oauth2provider.base
 
+import com.yetu.oauth2provider.base.DefaultTestVariables
 import com.yetu.oauth2provider.oauth2.models.YetuUser
 import com.yetu.oauth2provider.registry.TestRegistry
-import com.yetu.oauth2provider.testdata.DefaultTestVariables
 import com.yetu.oauth2provider.utils.Config
 import play.api.Logger
