--- test/com/yetu/oauth2provider/base/TestTags.scala
@@ -1,5 +1,0 @@
-package com.yetu.oauth2provider.base
-
-import org.scalatest.Tag
-
-object TempTag extends Tag("Temp")
