+++ test/com/yetu/oauth2provider/services/data/ClientServiceSpec.scala
@@ -1,5 +1,6 @@
 package com.yetu.oauth2provider.services.data
 
+import com.yetu.oauth2provider.base.DataServiceBaseSpec
 import com.yetu.oauth2provider.oauth2.models.OAuth2Client
 import com.yetu.oauth2provider.registry.{ TestRegistry, IntegrationTestRegistry }
