+++ test/com/yetu/oauth2provider/services/data/PermissionsServiceSpec.scala
@@ -1,5 +1,6 @@
 package com.yetu.oauth2provider.services.data
 
+import com.yetu.oauth2provider.base.DataServiceBaseSpec
 import com.yetu.oauth2provider.registry.{ TestRegistry, IntegrationTestRegistry }
 
 //TODO: implement permissions correctly and make sure this test leaves no traces behind.
