+++ test/com/yetu/oauth2provider/services/data/UserContactDataServiceSpec.scala
@@ -1,5 +1,6 @@
 package com.yetu.oauth2provider.services.data
 
+import com.yetu.oauth2provider.base.DataServiceBaseSpec
 import com.yetu.oauth2provider.models.DataUpdateRequest
 import com.yetu.oauth2provider.registry.{ TestRegistry, IntegrationTestRegistry }
 import com.yetu.oauth2provider.services.data.iface.IPersonService
