+++ test/com/yetu/oauth2provider/services/data/UserServiceSpec.scala
@@ -1,5 +1,6 @@
 package com.yetu.oauth2provider.services.data
 
+import com.yetu.oauth2provider.base.DataServiceBaseSpec
 import com.yetu.oauth2provider.registry.{ TestRegistry, IntegrationTestRegistry }
 import com.yetu.oauth2provider.services.data.iface.IPersonService
 import com.yetu.oauth2provider.utils.DateUtility._
